Made attributes a computer property of TextStyle instead of an internal function to the framework
